Refactor: ExpenseList와 관련된 BindingAdapter 이동
