+++ app/src/main/java/com/kappzzang/jeongsan/util/BindingAdapter.kt
@@ -68,17 +68,4 @@ object BindingAdapter {
         }
     }
 
-    @BindingAdapter("expenseItems")
-    @JvmStatic
-    fun attachExpenseList(
-        recyclerView: RecyclerView,
-        items: StateFlow<com.kappzzang.jeongsan.expenselist.ExpenseListViewUIData>?
-    ) {
-        items?.let {
-            (recyclerView.adapter as? com.kappzzang.jeongsan.expenselist.ExpenseListAdapter)
-                ?.submitList(
-                    it.value.expenseItems
-                )
-        }
-    }
 }
